requirements/_base.txt
asgiref==3.6.0
coverage==4.5.3
Django==3.2.17
djangorestframework==3.14.0
Pillow==9.4.0
psycopg2-binary==2.8.3
pytz==2019.2
sqlparse==0.3.0
